+++ CHANGES.rst
@@ -2,6 +2,12 @@
 Release Notes
 =============
 
+Version 1.0.1 (2017.03.03)
+==========================
+- show help message if no package name was specified for "create" command
+- remove unused 'timeout' argument from test_envs.py
+- reorganize if-statement in "list" command to display "Invalid request" when appropriate
+
 Version 1.0.0 (2017.03.02)
 ==========================
 - separate **install**, **uninstall**, **create** and **list** functions into different modules